+++ DB/DB_WhitelistedItems.lua
@@ -1,0 +1,7 @@
+local addon, addonTbl = ...;
+local L = addonTbl.L;
+
+local whitelistedItems = {
+};
+
+addonTbl.whitelistedItems = whitelistedItems;