+++ handlers/FunctionHandler.lua
@@ -20,145 +20,6 @@ addonTbl.GetCount = function(tbl, itemID)
 end
 -- Synopsis: Used to count records in a table or how many times an item has been seen by the player.
 
-addonTbl.Remove = function(arg)
-	if tonumber(arg) then -- The passed argument is a number or item ID.
-		arg = tonumber(arg);
-		if LastSeenClassicItemsDB[arg] then
-			if LastSeenClassicItemsDB[arg].itemLink ~= nil then
-				print(L["ADDON_NAME"] .. LastSeenClassicItemsDB[arg].itemLink .. L["INFO_MSG_ITEM_REMOVED"]);
-			else
-				print(L["ADDON_NAME"] .. arg .. L["INFO_MSG_ITEM_REMOVED"]);
-			end
-			LastSeenClassicItemsDB[arg] = nil;
-		end
-	elseif not tonumber(arg) then -- The passed argument isn't a number, and is likely an item's link.
-		arg = (GetItemInfoInstant(arg)); -- Converts the supposed item link into an item ID.
-		if tonumber(arg) then
-			arg = tonumber(arg);
-			if LastSeenClassicItemsDB[arg] then
-				if LastSeenClassicItemsDB[arg].itemLink ~= nil then
-					print(L["ADDON_NAME"] .. LastSeenClassicItemsDB[arg].itemLink .. L["INFO_MSG_ITEM_REMOVED"]);
-				else
-					print(L["ADDON_NAME"] .. arg .. L["INFO_MSG_ITEM_REMOVED"]);
-				end
-				LastSeenClassicItemsDB[arg] = nil;
-			end
-		end
-	else
-		print(L["ADDON_NAME"] .. L["ERROR_MSG_BAD_REQUEST"]);
-	end
-	
-	if (LastSeenClassicLootTemplate[arg]) then LastSeenClassicLootTemplate[arg] = nil end; -- Remove all associated entries that the player looted the item from.
-end
--- Synopsis: Allows the player to remove undesired items from the items table using its ID or link.
-
-addonTbl.Search = function(query)
-	local itemsFound = 0;
-	local questsFound = 0;
-	local queryType = string.sub(query, 1, 1);
-	--local query = string.match(query, queryType .. "%s" .. "(.*)");
-	if tonumber(query) ~= nil then -- It's an ID
-		query = tonumber(query);
-		if LastSeenItemsDB[query] then
-			print(query .. ": " .. LastSeenItemsDB[query].itemLink .. " (" .. addonTbl.GetCount(LastSeenLootTemplate, query) .. ") | " .. LastSeenItemsDB[query].lootDate .. " | " .. LastSeenItemsDB[query].source .. " | " ..
-			LastSeenItemsDB[query].location);
-			itemsFound = itemsFound + 1;
-		end
-	else
-		for k, v in pairs(LastSeenItemsDB) do
-			if v.source ~= L["INFO_MSG_MISCELLANEOUS"] or v.source or v.location or v.itemLink then
-				if string.find(string.lower(v.itemLink), string.lower(query)) then
-					local itemID = (GetItemInfoInstant(k));
-					print(k .. ": " .. v.itemLink .. " (" .. addonTbl.GetCount(LastSeenLootTemplate, itemID) .. ") | " .. v.lootDate .. " | " .. v.source .. " | " .. v.location);
-					itemsFound = itemsFound + 1;
-				end
-				if string.find(string.lower(v.source), string.lower(query)) then
-					local itemID = (GetItemInfoInstant(k));
-					print(k .. ": " .. v.itemLink .. " (" .. addonTbl.GetCount(LastSeenLootTemplate, itemID) .. ") | " .. v.lootDate .. " | " .. v.source .. " | " .. v.location);
-					itemsFound = itemsFound + 1;
-				end
-				if string.find(string.lower(v.location), string.lower(query)) then
-					local itemID = (GetItemInfoInstant(k));
-					if v.itemLink == "" then
-						print(k .. ": " .. v.itemName .. " (" .. addonTbl.GetCount(LastSeenLootTemplate, itemID) .. ") | " .. v.lootDate .. " | " .. v.source .. " | " .. v.location);
-					else
-						if v.lootDate == nil then
-							--
-						else
-							print(k .. ": " .. v.itemLink .. " (" .. addonTbl.GetCount(LastSeenLootTemplate, itemID) .. ") | " .. v.lootDate .. " | " .. v.source .. " | " .. v.location);
-						end
-					end
-					itemsFound = itemsFound + 1;
-				end
-			end
-		end
-		if itemsFound == 0 then
-			print(L["ADDON_NAME"] .. L["ERROR_MSG_NO_ITEMS_FOUND"] .. " (" .. query .. ")");
-		else
-			print(L["ADDON_NAME"] .. itemsFound .. L["INFO_MSG_RESULTS"]);
-		end
-	end
-end
---[[
-	Synopsis: Allows the player to search the items table by an item's ID or its partial/full name, a creature's name, or a zone's name.
-	Use Case(s):
-		- Item: The most common search is by an item's ID or partial/full name to provide another player with proof an item still drops.
-		- Creature: A search to see what items a particular creature has dropped. This is made possible by keeping track of the items a creature has dropped (call a "loot table").
-		- Zone: An entire zone search, doesn't matter which creature in the zone dropped the item. Provides the player with every item that dropped in the provided zone. Partial/full names supported.
-]]
-
-addonTbl.Manual = function(args)
-	if args == "" then
-		print(L["ADDON_NAME"]);
-		print("Commands: " .. L["CMD_DISCORD"] .. ", " .. L["CMD_HISTORY"] .. ", " .. ", " .. L["CMD_MAN"] .. ", " .. L["CMD_REMOVE"] .. ", " .. L["CMD_SEARCH"] .. ", " .. L["CMD_VIEW"]);
-	elseif args == L["CMD_DISCORD"] then
-		print(L["ADDON_NAME"]);
-		print("Command: " .. L["CMD_DISCORD"]);
-		print("Description: Provides a link in the chat frame to the Discord server used by the addon. It's the server that belongs to the infamous ALL THE THINGS addon.");
-	elseif args == L["CMD_HISTORY"] then
-		print(L["ADDON_NAME"]);
-		print("Command: " .. L["CMD_HISTORY"]);
-		print("Description: Addon output to the chat is ephemeral, and is unavailable should the player disconnect or log off. As such, the history command can be used to review the last 20 items seen by the player's characters. This is account wide and applies to all sources.");
-	elseif args == L["CMD_MAN"] then
-		print(L["ADDON_NAME"]);
-		print("Command: " .. L["CMD_MAN"]);
-		print("Description: A manual used to learn about commands.");
-	elseif args == L["CMD_REMOVE"] then
-		print(L["ADDON_NAME"]);
-		print("Command: " .. L["CMD_REMOVE"]);
-		print("Description: This will remove items from the items table. Both item ID and item links are supported.");
-		print("Aliases: " .. L["CMD_REMOVE_SHORT"]);
-	elseif args == L["CMD_SEARCH"] then
-		print(L["ADDON_NAME"]);
-		print("Command: " .. L["CMD_SEARCH"]);
-		print("Description: Allows the player to search the items table by an item ID, name, or link. Search queries will query items, creatures, and zones automatically. Be as specific or ambiguous as you want.");
-	elseif args == L["CMD_VIEW"] then
-		print(L["ADDON_NAME"]);
-		print("Command: " .. L["CMD_VIEW"]);
-		print("Description: When corrupted items are discovered, they're automatically removed when the addon is loaded. This command allows the player to see a list of that which was removed.");
-	end
-end
-
-addonTbl.GetCurrentMap = function()
-	local uiMapID = C_Map.GetBestMapForUnit("player");
-	local isInInstance;
-	
-	if uiMapID then -- A map ID was found and is usable.
-		local uiMap = C_Map.GetMapInfo(uiMapID);
-		if not uiMap.mapID then return end;
-		if not LastSeenClassicMapsDB[uiMap.mapID] then
-			LastSeenClassicMapsDB[uiMap.mapID] = uiMap.name;
-		end
-
-		addonTbl.currentMap = uiMap.name;
-	else
-		C_Timer.After(3, addonTbl.GetCurrentMap); -- Recursively call the function every 3 seconds until a map ID is found.
-	end
-	
-	return addonTbl.currentMap;
-end
--- Synopsis: Gets the player's current map so an item can be accurately recorded.
-
 addonTbl.DataIsValid = function(itemID)
 	if itemID == nil then
 		return false;
@@ -180,39 +41,14 @@ end
 	Written by: Arcanemagus
 ]]
 
-addonTbl.OnTooltipSetItem = function(tooltip)
-	local isIgnored = false;
-	local _, itemLink = tooltip:GetItem();
-	if not itemLink then return end;
-
-	local itemID = (GetItemInfoInstant(itemLink)); if not itemID then return end; -- To handle reagents in the tradeskill window.
-
-	local itemTypeID = select(12, GetItemInfo(itemID));
-
-	if LastSeenClassicItemsDB[itemID] then -- Item exists in the database; therefore, show its data.
-		local frame, text;
-		for i = 1, 30 do
-			frame = _G[tooltip:GetName() .. "TextLeft" .. i]
-			if frame then text = frame:GetText() end;
-			if text and string.find(text, "LastSeen") then return end;
-		end
-		if addonTbl.DataIsValid(itemID) then
-			tooltip:AppendText(" (|cffadd8e6" .. LastSeenClassicItemsDB[itemID].source .. "|r)");
-			tooltip:AddLine(L["ADDON_NAME"] .. "|cffadd8e6" .. LastSeenClassicItemsDB[itemID].location .. "|r | |cffadd8e6" .. LastSeenClassicItemsDB[itemID].lootDate .. "|r");
-			tooltip:Show();
-		end
-	end
-	
-	if addonTbl.Contains(addonTbl.ignoredItemCategories, nil, "itemType", select(6, GetItemInfo(itemID))) then isIgnored = true end;
-	if not isIgnored then if addonTbl.Contains(addonTbl.ignoredItemCategories, nil, "itemType", select(7, GetItemInfo(itemID))) then isIgnored = true end end;
-	if not isIgnored then if addonTbl.Contains(addonTbl.ignoredItemCategories, nil, "itemType", select(9, GetItemInfo(itemID))) then isIgnored = true end end;
-	
-	if isIgnored then
-		tooltip:AddLine("\n" .. L["ADDON_NAME"] .. "|cffffffff" .. L["INFO_MSG_IGNORED_ITEM"] .. "|r");
-		tooltip:Show();
-	end
+addonTbl.InitializeSavedVars = function()
+	if LastSeenClassicMapsDB == nil then LastSeenClassicMapsDB = {} end;
+	if LastSeenClassicCreaturesDB == nil then LastSeenClassicCreaturesDB = {} end;
+	if LastSeenClassicItemsDB == nil then LastSeenClassicItemsDB = {} end;
+	if LastSeenClassicSettingsCacheDB == nil then LastSeenClassicSettingsCacheDB = {} end;
+	if LastSeenClassicLootTemplate == nil then LastSeenClassicLootTemplate = {} end;
+	if LastSeenClassicHistoryDB == nil then LastSeenClassicHistoryDB = {} end;
 end
--- Synopsis: Adds text to the tooltip regarding the source of an item, the location in which the player was when the item was looted, and the date it was looted.
 
 addonTbl.ExtractItemLink = function(constant)
 	local extractedLink, itemID, itemLink;
@@ -292,6 +128,19 @@ addonTbl.RollHistory = function()
 end
 -- Synopsis: Maintains the history table, to always keep it at the maximum number of entries, which is currently 20.
 
+addonTbl.DateFormat = function(format)
+	for k, v in pairs(LastSeenClassicItemsDB) do
+		if tonumber(format) then -- The player passed in a number so set the format to DAY/MONTH/YEAR.
+			local month, day, year = string.match(LastSeenClassicItemsDB[k]["lootDate"], "^(%d%d)/(%d%d)/(%d%d%d%d)$");
+			LastSeenClassicItemsDB[k]["lootDate"] = string.format("%s/%s/%s", day, month, year);
+		else
+			local day, month, year = string.match(LastSeenClassicItemsDB[k]["lootDate"], "^(%d%d)/(%d%d)/(%d%d%d%d)$");
+			LastSeenClassicItemsDB[k]["lootDate"] = string.format("%s/%s/%s", month, day, year);
+		end
+	end
+end
+-- Synopsis: Changes the date format for existing items from MONTH/DAY/YEAR to DAY/MONTH/YEAR or vice versa.
+
 addonTbl.GetItemInfo = function(itemLink, slot)
 	local lootSources = { GetLootSourceInfo(slot) };
 
@@ -301,38 +150,39 @@ addonTbl.GetItemInfo = function(itemLink, slot)
 		for j = 1, #lootSources, 2 do
 			if itemLink then
 				itemName = GetItemInfo(itemLink);
+				itemRarity = select(3, GetItemInfo(itemLink));
 				itemID, itemType, itemSubType, itemEquipLoc, itemIcon = GetItemInfoInstant(itemLink);
 				local type, _, _, _, _, creatureID = strsplit("-", lootSources[j]);
 				if itemID then -- To catch items without an item ID.
 					addonTbl.itemsToSource[itemID] = tonumber(creatureID);
 					addonTbl.itemSourceCreatureID = addonTbl.itemsToSource[itemID];
 					
-					if addonTbl.Contains(addonTbl.ignoredItemCategories, nil, "itemType", itemType) then return end;
-					if addonTbl.Contains(addonTbl.ignoredItemCategories, nil, "itemType", itemSubType) then return end;
-					if addonTbl.Contains(addonTbl.ignoredItemCategories, nil, "itemType", itemEquipLoc) then return end;
-					if addonTbl.Contains(addonTbl.ignoredItems, itemID, nil, nil) then return end;
-					
-					if LastSeenClassicItemsDB[itemID] then -- Item seen again.
-						if LastSeenClassicCreaturesDB[addonTbl.itemSourceCreatureID] then
-							addonTbl.AddItem(itemID, itemLink, itemName, itemRarity, itemType, itemSubType, itemEquipLoc, itemIcon, L["DATE"], addonTbl.currentMap, "Creature", LastSeenClassicCreaturesDB[addonTbl.itemSourceCreatureID].unitName, "Update");
-						elseif addonTbl.encounterID then
-							addonTbl.AddItem(itemID, itemLink, itemName, itemRarity, itemType, itemSubType, itemEquipLoc, itemIcon, L["DATE"], addonTbl.currentMap, "Encounter", LastSeenClassicEncountersDB[addonTbl.encounterID], "Update");
-						elseif addonTbl.target then
-							addonTbl.AddItem(itemID, itemLink, itemName, itemRarity, itemType, itemSubType, itemEquipLoc, itemIcon, L["DATE"], addonTbl.currentMap, "Object", addonTbl.target, "Update");
-						else
-							if addonTbl.mode ~= L["QUIET_MODE"] then print(L["ADDON_NAME"] .. itemLink .. L["ERROR_MSG_UNKNOWN_SOURCE"]) end;
-							addonTbl.AddItem(itemID, itemLink, itemName, itemRarity, itemType, itemSubType, itemEquipLoc, itemIcon, L["DATE"], addonTbl.currentMap, "Miscellaneous", L["INFO_MSG_MISCELLANEOUS"], "Update");
-						end
-					else -- Item seen for first time.
-						if LastSeenClassicCreaturesDB[addonTbl.itemSourceCreatureID] then
-							addonTbl.AddItem(itemID, itemLink, itemName, itemRarity, itemType, itemSubType, itemEquipLoc, itemIcon, L["DATE"], addonTbl.currentMap, "Creature", LastSeenClassicCreaturesDB[addonTbl.itemSourceCreatureID].unitName, "New");
-						elseif addonTbl.encounterID then
-							addonTbl.AddItem(itemID, itemLink, itemName, itemRarity, itemType, itemSubType, itemEquipLoc, itemIcon, L["DATE"], addonTbl.currentMap, "Encounter", LastSeenClassicEncountersDB[addonTbl.encounterID], "New");
-						elseif addonTbl.target then
-							addonTbl.AddItem(itemID, itemLink, itemName, itemRarity, itemType, itemSubType, itemEquipLoc, itemIcon, L["DATE"], addonTbl.currentMap, "Object", addonTbl.target, "New");
-						else
-							if addonTbl.mode ~= L["QUIET_MODE"] then print(L["ADDON_NAME"] .. itemLink .. L["ERROR_MSG_UNKNOWN_SOURCE"]) end;
-							addonTbl.AddItem(itemID, itemLink, itemName, itemRarity, itemType, itemSubType, itemEquipLoc, itemIcon, L["DATE"], addonTbl.currentMap, "Miscellaneous", L["INFO_MSG_MISCELLANEOUS"], "New");
+					if itemRarity >= addonTbl.rarity then
+						if addonTbl.Contains(addonTbl.whitelistedItems, itemID, nil, nil) then
+							-- Continue
+						elseif addonTbl.Contains(addonTbl.ignoredItemCategories, nil, "itemType", itemType) then return;
+						elseif addonTbl.Contains(addonTbl.ignoredItemCategories, nil, "itemType", itemSubType) then return;
+						elseif addonTbl.Contains(addonTbl.ignoredItemCategories, nil, "itemType", itemEquipLoc) then return;
+						elseif addonTbl.Contains(addonTbl.ignoredItems, itemID, nil, nil) then return end;
+						
+						if LastSeenClassicItemsDB[itemID] then -- Item seen again.
+							if LastSeenClassicCreaturesDB[addonTbl.itemSourceCreatureID] then
+								addonTbl.AddItem(itemID, itemLink, itemName, itemRarity, itemType, itemSubType, itemEquipLoc, itemIcon, L["DATE"], addonTbl.currentMap, "Creature", LastSeenClassicCreaturesDB[addonTbl.itemSourceCreatureID].unitName, "Update");
+							elseif addonTbl.target ~= "" then
+								addonTbl.AddItem(itemID, itemLink, itemName, itemRarity, itemType, itemSubType, itemEquipLoc, itemIcon, L["DATE"], addonTbl.currentMap, "Object", addonTbl.target, "Update");
+							else
+								if addonTbl.mode ~= GM_SURVEY_NOT_APPLICABLE then print(L["ADDON_NAME"] .. itemLink .. L["ERROR_MSG_UNKNOWN_SOURCE"]) end;
+								addonTbl.AddItem(itemID, itemLink, itemName, itemRarity, itemType, itemSubType, itemEquipLoc, itemIcon, L["DATE"], addonTbl.currentMap, "Miscellaneous", L["INFO_MSG_MISCELLANEOUS"], "Update");
+							end
+						else -- Item seen for first time.
+							if LastSeenClassicCreaturesDB[addonTbl.itemSourceCreatureID] then
+								addonTbl.AddItem(itemID, itemLink, itemName, itemRarity, itemType, itemSubType, itemEquipLoc, itemIcon, L["DATE"], addonTbl.currentMap, "Creature", LastSeenClassicCreaturesDB[addonTbl.itemSourceCreatureID].unitName, "New");
+							elseif addonTbl.target ~= "" then
+								addonTbl.AddItem(itemID, itemLink, itemName, itemRarity, itemType, itemSubType, itemEquipLoc, itemIcon, L["DATE"], addonTbl.currentMap, "Object", addonTbl.target, "New");
+							else
+								if addonTbl.mode ~= GM_SURVEY_NOT_APPLICABLE then print(L["ADDON_NAME"] .. itemLink .. L["ERROR_MSG_UNKNOWN_SOURCE"]) end;
+								addonTbl.AddItem(itemID, itemLink, itemName, itemRarity, itemType, itemSubType, itemEquipLoc, itemIcon, L["DATE"], addonTbl.currentMap, "Miscellaneous", L["INFO_MSG_MISCELLANEOUS"], "New");
+							end
 						end
 					end
 				end